requirements.txt
torch==2.2.1
transformers==4.39.3
datasets==2.18.0
accelerate==0.28.0
wandb==0.16.6
scipy==1.12.0
numpy==1.26.4
pyyaml==6.0.1
tqdm==4.66.2
